Adding JSON & Tuples Instructions
The data source is now a JSON file. The instructions are currently hard coded in the caller. I have created a Tuple List combining action and value. This allows "instructions" to be emulated, and the code can print the numbers if they are numbers and select the command buttons if the action is anything other than a number (e.g. total). Next we can look to add audio to each number so it can sound like the boss is dictating what numbers to add and whether to total or subtotal etc.